refactor: replace TestRPC with RpcSpy for improved RPC call testing
